Calculate ball speed and try to extrapolate it
The ball speed is calculated using two consecutives frames. If at some
point the ball is lost, its position will be extrapolated using the last
known speed.

Please note that using two consecutives frames is highly susceptible to
noise in the input data. Another way will need to be used to get a speed
that is useable to predict the ball's position.